Move 'pin' package inside 'security'
