+++ app/src/main/java/org/simple/clinic/di/AppModule.kt
@@ -12,7 +12,7 @@ import org.simple.clinic.AppDatabase
 import org.simple.clinic.crash.CrashReporterModule
 import org.simple.clinic.login.LoginModule
 import org.simple.clinic.patient.fuzzy.AgeFuzzerModule
-import org.simple.clinic.pin.BruteForceProtectionModule
+import org.simple.clinic.security.pin.BruteForceProtectionModule
 import org.simple.clinic.qrscan.QrModule
 import org.simple.clinic.registration.RegistrationModule
 import org.simple.clinic.storage.Migration_10_11
