+++ app/src/main/java/org/simple/clinic/login/applock/AppLockScreen.kt
@@ -14,7 +14,7 @@ import kotterknife.bindView
 import org.simple.clinic.R
 import org.simple.clinic.activity.TheActivity
 import org.simple.clinic.facility.change.FacilityChangeScreen
-import org.simple.clinic.pin.PinEntryCardView
+import org.simple.clinic.security.pin.PinEntryCardView
 import org.simple.clinic.router.screen.BackPressInterceptCallback
 import org.simple.clinic.router.screen.BackPressInterceptor
 import org.simple.clinic.router.screen.ScreenRouter
