+++ app/src/main/java/org/simple/clinic/security/pin/BruteForceProtectionConfig.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.pin
+package org.simple.clinic.security.pin
 
 import org.threeten.bp.Duration
