+++ app/src/main/java/org/simple/clinic/security/pin/PinEntryCardViewEvents.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.pin
+package org.simple.clinic.security.pin
 
 import org.simple.clinic.widgets.UiEvent
